yaml session loader
