+++ quickfixj-core/src/test/java/quickfix/CachedFileStoreTest.java
@@ -35,7 +35,7 @@ public class CachedFileStoreTest extends AbstractMessageStoreTest {
     }
 
     protected MessageStoreFactory getMessageStoreFactory() throws ConfigError, FieldConvertError {
-        SessionSettings settings = new SessionSettings(getConfigurationFileName());
+        SessionSettings settings =  SessionSettingsLoaders.load(getConfigurationFileName());
         // Initialize the session settings from the defaults
         settings.setString(getSessionID(), FileStoreFactory.SETTING_FILE_STORE_PATH, settings
                 .getString(FileStoreFactory.SETTING_FILE_STORE_PATH));
