+++ quickfixj-core/src/test/java/quickfix/FileStoreLowMemTest.java
@@ -22,7 +22,7 @@ package quickfix;
 public class FileStoreLowMemTest extends FileStoreTest {
     @Override
     protected MessageStoreFactory getMessageStoreFactory() throws ConfigError, FieldConvertError {
-        SessionSettings settings = new SessionSettings(getConfigurationFileName());
+        SessionSettings settings = SessionSettingsLoaders.load(getConfigurationFileName());
         // Initialize the session settings from the defaults
         settings.setString(getSessionID(), FileStoreFactory.SETTING_FILE_STORE_PATH, settings
                 .getString(FileStoreFactory.SETTING_FILE_STORE_PATH));
