+++ quickfixj-core/src/test/java/quickfix/FileStoreTest.java
@@ -37,7 +37,7 @@ public class FileStoreTest extends AbstractMessageStoreTest {
     }
 
     protected MessageStoreFactory getMessageStoreFactory() throws ConfigError, FieldConvertError {
-        SessionSettings settings = new SessionSettings(getConfigurationFileName());
+        SessionSettings settings = SessionSettingsLoaders.load(getConfigurationFileName());
         // Initialize the session settings from the defaults
         settings.setString(getSessionID(), FileStoreFactory.SETTING_FILE_STORE_PATH, settings
                 .getString(FileStoreFactory.SETTING_FILE_STORE_PATH));
